Move ICommand to interfaces
