+++ src/Application/Commands/Commands.cs
@@ -2,7 +2,7 @@ namespace Application.Commands;
 
 public class Commands
 {
-  public static readonly Application.Commands.ICommand[] AllowedCommands =
+  public static readonly Application.Interfaces.ICommand[] AllowedCommands =
   {
     new LoginCommand(),
     new HelpCommand(),
